requirements.txt
discord.py==2.1.0
transformers==4.27.1
torch==1.13.1
requests==2.28.1
python-dotenv==0.21.0
